Consistent code style was applied by prettier package
Actual formatting was made with three commands:
node ./node_modules/.bin/prettier --write 'contracts/**/*.sol'
node ./node_modules/.bin/prettier --write 'test/**/*.sol'
node ./node_modules/.bin/prettier --write 'test/**/*.js'

+++ test/mocks/MockedStakingToken.sol
@@ -3,8 +3,7 @@ pragma solidity >=0.7.0 <0.8.0;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract MockedStakingToken is ERC20 {
-
-    uint private constant TOTAL_SUPPLY = 10000 * (10**18);
+    uint256 private constant TOTAL_SUPPLY = 10000 * (10**18);
 
     constructor() ERC20("Nominex LP", "NMXLP") {
         _mint(msg.sender, TOTAL_SUPPLY);
@@ -13,5 +12,4 @@ contract MockedStakingToken is ERC20 {
     function token0() external view returns (address) {
         return address(this);
     }
-
 }
